% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{build_dataset_from_fl_model}
\alias{build_dataset_from_fl_model}
\title{build_dataset_from_fl_model}
\usage{
build_dataset_from_fl_model(data, data_specification, first_level_model,
  num_of_model, transformation_function)
}
\description{
Builds a dataset using the first level model and the transformation.
}
