% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{extract_model_data}
\alias{extract_model_data}
\title{extract_model_data}
\usage{
extract_model_data(data_modeled, num_of_model)
}
\description{
Extracts all of the \code{num_of_model}-th distribution fits for all instances.
}
