% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{get_instance_measurements}
\alias{get_instance_measurements}
\title{get_instance_measurements}
\usage{
get_instance_measurements(instance_id, data_all, cols_ids, cols_measurements)
}
\description{
Retrieves measurements of all attributes of an instance. One row denotes one event.
}
