% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_measurement_colnames}
\alias{get_measurement_colnames}
\title{get_measurement_colnames}
\usage{
get_measurement_colnames(data)
}
\description{
Get column names of measurements (without suffix) from data.frame in
event-instance format.
}
