% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{get_object_measurements}
\alias{get_object_measurements}
\title{get_object_measurements}
\usage{
get_object_measurements(object_id, data_all, cols_ids, cols_measurements)
}
\description{
Retrieves measurements of all attributes of an object. One row contains data with measurements
for a single match.
}
