% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_static_colnames}
\alias{get_static_colnames}
\title{get_id_colnames}
\usage{
get_static_colnames(data)
}
\description{
Get static columns (i.e. ones that are not modelled) from data.frame in
event-instance format. For example: y, TIME, ID ...
}
