% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{get_transformation}
\alias{get_transformation}
\title{get_transformation}
\usage{
get_transformation(transformation)
}
\description{
Returns a function representing the transformation (i.e. way of obtaining
feature vectors). Returns the function itself if a function
is passed in as parameter. If a string is passed, it has to be one of the
supported transformations (currently only "means").
}
