% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_model_mvnormal_iw.R
\name{input_model_mvnormal_iw}
\alias{input_model_mvnormal_iw}
\title{input_model_mvnormal_iw}
\usage{
input_model_mvnormal_iw(data, num_models, prior)
}
\arguments{
\item{data}{Data frame of events with attributes to model jointly.}

\item{num_models}{Number of distributions to fit.}

\item{prior}{List of lists containing priors for mean and covariance matrix. Should
contain entries 'mean' and 'sigma'. Inner list corresponding to entry 'mean'
should specify parameters for the multivariate normal prior of the mean.
It's entries should be 'mean' (vector) and 'sigma' (matrix) of the appropriate size.
List corresponding to the outer list key 'sigma' should specify parameters for the Inverse Wishart
prior of the covariance matrix. Keys in this list should be 'a' (scalar)
and 'S' (matrix). For example:
\code{list(mean = list(mean = .., sigma = ..), sigma = list(a = .., S = ..))},
values are omitted for clarity.}
}
\value{
List of \code{num_models} S3 classes of type 'mvnormal'.
  The class 'mvnormal' has the following methods defined: \code{mean} and
  \code{sample}. These both sample from the posterior distribution of the mean
  of the multivariate distribution fit.
}
\description{
Multivariate Normal - Inverse Wishart model. Fits \code{num_models} Multivariate
Normal distributions to a data frame according to Multivariate Normal - Inverse Wishart model.
If prior is not passed to the function, classic Multivariate Normal model with sample covariance
is used.
}
\seealso{
Other input models: \code{\link{input_model_bernoulli}},
  \code{\link{input_model_normal_ig}},
  \code{\link{input_model_normal}},
  \code{\link{input_model_poisson}}
}
