% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_model_normal.R
\name{input_model_normal}
\alias{input_model_normal}
\title{input_model_normal}
\usage{
input_model_normal(data, num_models, prior)
}
\arguments{
\item{data}{Numerical vector of data points.}

\item{num_models}{Number of distributions to fit.}

\item{prior}{List containing the parameters of the Normal prior distribution of
the mean. List has to contain entries "mu" and "var" being the mean and
variance parameters. For example: \code{list(mu = 10, var = 2)}.}
}
\value{
List of \code{num_models} S3 classes of type 'normal'.
  The class 'normal' has the following methods defined: mean, sample and var.
}
\description{
Normal - sample variance model. Fits \code{num_models} normal models with
variance being sample variance to a numerical vector (attribute).
}
\seealso{
Other input models: \code{\link{input_model_bernoulli}},
  \code{\link{input_model_mvnormal_iw}},
  \code{\link{input_model_normal_ig}},
  \code{\link{input_model_poisson}}
}
