% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{model_first_level}
\alias{model_first_level}
\title{model_first_level}
\usage{
model_first_level(data_to_model, data_all, input_model_specification,
  num_models, weighting = NULL, priors = NULL, data_specification)
}
\description{
Builds a first level model for objects in the data set. Result is a list of
object models.
}
