% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{model_first_level_object}
\alias{model_first_level_object}
\title{model_first_level_object}
\usage{
model_first_level_object(data, input_model_specification, num_models,
  prior = NULL)
}
\description{
Model the attributes of an object according to the input specification.
}
