% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{model_inputs}
\alias{model_inputs}
\title{model_inputs}
\usage{
model_inputs(data_to_model, data_all, input_model_specification, num_models,
  weighting = NULL, priors = NULL, data_specification)
}
\description{
Samples distributions for each attribute of the instances. Modularized as
function to prevent unnecessary repetition, since the same procedure is used
in both \code{\link{match_forecast_model}} and
\code{\link{predict.match_forecast_model}}.
}
