% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_forecast.R
\name{model_inputs_row}
\alias{model_inputs_row}
\title{model_inputs_row}
\usage{
model_inputs_row(data, input_model_specification, num_models, prior = NULL)
}
\description{
Model the attributes of an instance (row) according to the input specification.
}
