% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{reformat_to_df}
\alias{reformat_to_df}
\title{reformat_to_df}
\usage{
reformat_to_df(data_modeled, data_specification)
}
\description{
Sets column names of the newly created data frame according to the package
structure. Also sets column types.
}
