% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(row_transformation)
}
\arguments{
\item{row_transformation}{transformation to apply to each instance's model.}
}
\description{
Transforms distributions fitted to attributes to feature vectors by applying
\code{row_transformation}. \code{row_transformation} acts on a single instance.
}
