% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transformation_means}
\alias{transformation_means}
\title{transformation_means}
\usage{
transformation_means(instance_model)
}
\description{
Extracts means from all the distributions fits to the attributes of an instance.
}
