% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighting_gaussian}
\alias{weighting_gaussian}
\title{weighting_gaussian}
\usage{
weighting_gaussian(data, sd)
}
\description{
TODO: decide whether to include this or not. Draws a sample from data using gaussian weighting.
}
