% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_model_normal.R, R/input_model_normal_ig.R
\name{input_model_normal_ig}
\alias{input_model_normal_ig}
\alias{input_model_normal_ig}
\title{input_model_normal_ig}
\usage{
input_model_normal_ig(data, num_models, prior)

input_model_normal_ig(data, num_models, prior)
}
\arguments{
\item{data}{Numerical vector of data points.}

\item{num_models}{Number of distributions to fit.}

\item{prior}{List of two lists with keys 'mean' and 'var'. List corresponding to key
'mean' needs to have entries 'mu' (mean of the prior mean distribution) and 'k'
(variance scaler). List corresponding to key 'var' needs to have entries 'a' and 'b'
corresponding to rate and scale of the inverse gamma prior for variance. For example:
\code{list(mean = list(mu = 10, k = 2), var = list(a = 5, b = 2))}.}

\item{prior}{List containing the parameters for Normal prior distribution of the mean
and Inverse Gamma prior distribution for the variance. For example:
\code{list(mean = list(mu = 10, k = 2), var = list(a = 5, b = 3)}.}
}
\value{
List of \code{num_samples} S3 classes of type 'normal'.
  The class 'normal' has the following methods defined: mean, sample and var.

List of \code{num_models} S3 classes of type 'normal'.
  The class 'normal' has the following methods defined: mean, sample and var.
}
\description{
Fits \code{num_samples} normal models according to normal - inverse gamma model.

Normal - Inverse Gamma model. Jointly models mean and variance of a normal distribution.
}
\seealso{
Other input models: \code{\link{input_model_bernoulli}},
  \code{\link{input_model_mvnormal_iw}},
  \code{\link{input_model_normal}},
  \code{\link{input_model_poisson}}

Other input models: \code{\link{input_model_bernoulli}},
  \code{\link{input_model_mvnormal_iw}},
  \code{\link{input_model_normal}},
  \code{\link{input_model_poisson}}
}
